refactor: Clean up dropdown items
