+++ src/components/ControlContainer/_control-container.scss
@@ -18,4 +18,8 @@
   .container-title {
     padding-bottom: $spacing-09;
   }
+
+  .voice-info {
+    padding-bottom: $spacing-07;
+  }
 }
